First part of restructuring rgroup decomposition classes
