+++ Code/GraphMol/RGroupDecomposition/CMakeLists.txt
@@ -1,7 +1,8 @@
 
 remove_definitions(-DRDKIT_GRAPHMOL_BUILD)
 add_definitions(-DRDKIT_RGROUPDECOMPOSITION_BUILD)
-rdkit_library(RGroupDecomposition RGroupDecomp.cpp
+rdkit_library(RGroupDecomposition RGroupDecomp.cpp RGroupUtils.cpp
+				  RGroupDecompParams.cpp RGroupScore.cpp
   LINK_LIBRARIES
   FMCS ChemTransforms SubstructMatch SmilesParse 
  GraphMol RDGeneral)
