+++ Code/GraphMol/RGroupDecomposition/RGroupMatch.h
@@ -1,0 +1,18 @@
+#ifndef RGROUP_MATCH_DATA
+#define RGROUP_MATCH_DATA
+#include "RGroupData.h"
+
+namespace RDKit {
+typedef boost::shared_ptr<RGroupData> RData;
+typedef std::map<int, RData> R_DECOMP;
+struct RGroupMatch {
+  // RGroupMatch is the decomposition for a single molecule
+  size_t core_idx;   // index of the matching core
+  R_DECOMP rgroups;  // rlabel->RGroupData mapping
+
+  RGroupMatch(size_t core_index, R_DECOMP input_rgroups)
+      : core_idx(core_index), rgroups(std::move(input_rgroups)) {}
+};  
+
+}
+#endif
