+++ Code/GraphMol/RGroupDecomposition/RGroupScore.h
@@ -1,0 +1,52 @@
+#ifndef RGROUP_SCORE_H
+#define RGROUP_SCORE_H
+
+#include "RGroupMatch.h"
+#include <vector>
+#include <set>
+namespace RDKit {
+
+struct CartesianProduct {
+  std::vector<size_t> permutation;
+  std::vector<size_t> sizes;
+  size_t maxPermutations;
+  size_t permutationCount;
+  CartesianProduct(const std::vector<size_t> &inputSizes)
+      : permutation(inputSizes.size(), 0),
+        sizes(inputSizes),
+        permutationCount(0) {
+    maxPermutations = 1;
+    for (unsigned long size : sizes) {
+      maxPermutations *= size;  // may overflow....
+    }
+  }
+
+  bool next() {
+    ++permutationCount;
+    if (permutationCount == 1) {
+      return true;
+    }
+
+    return increment(0);
+  }
+
+  bool increment(size_t rowToIncrement) {
+    if (permutationCount > maxPermutations) {
+      return false;
+    }
+
+    permutation[rowToIncrement] += 1;
+    size_t max_index_of_row = sizes[rowToIncrement] - 1;
+    if (permutation[rowToIncrement] > max_index_of_row) {
+      permutation[rowToIncrement] = 0;
+      return increment(rowToIncrement + 1);
+    }
+    return true;
+  }
+};
+
+double score(const std::vector<size_t> &permutation,
+             const std::vector<std::vector<RGroupMatch>> &matches,
+             const std::set<int> &labels);
+}
+#endif
